refactor: split api-client into focused modules under lib/api
Extract response/request types, the request transport (sendRequest + token
cache + sessionStorage), and the per-domain clients (auth, outfits, users,
boards, admin) into separate files under lib/api/. lib/api-client.ts is now a
thin re-export shim so existing imports keep working unchanged.

Public API preserved exactly; tsc --noEmit and the Playwright smoke suite pass.
